pyproject.toml
[project]
name = "tiktok-save"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.6.1",
    "easyocr>=1.7.2",
    "faster-whisper>=1.2.1",
    "memory-profiler>=0.61.0",
    "openai>=2.13.0",
    "opencv-contrib-python>=4.13.0.90",
    "paddlepaddle>=3.3.0",
    "pillow>=12.1.0",
    "playwright==1.52.0",
    "rapidocr-onnxruntime>=1.4.4",
    "redis>=6.4.0",
    "tiktokapi==7.1.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]
